wpp: Merge preproc.c into wpp.c.
